Refactor journal application structure by removing main.py and introducing a modular design with separate files for CLI, database, and models. Implemented a new JournalDatabase class for data handling and updated CLI commands to interact with the database. Added initial test structure for future testing.
