+++ pyproject.toml
@@ -6,11 +6,13 @@ readme = "README.md"
 requires-python = ">=3.14"
 dependencies = [
     "pydantic>=2.12.4",
+    "sqlmodel>=0.0.27",
     "typer>=0.20.0",
 ]
 
 [dependency-groups]
 dev = [
+    "codeflash>=0.18.7",
     "pytest-cov>=7.0.0",
     "ruff>=0.14.6",
 ]
